Refactored & added mobile to profile, + types pack
