--- src/pages/common/Header/Menu/MenuMobile.tsx
@@ -1,18 +1,0 @@
-import React from 'react'
-import { NavLink } from 'react-router-dom'
-import { COMMUNITY_PAGES } from 'src/pages/PageList'
-import theme from 'src/themes/styled.theme'
-import { Flex } from 'rebass'
-
-interface IProps {
-  onClick?: () => void
-}
-
-export class MenuMobile extends React.Component<IProps> {
-  render() {
-    return <></>
-  }
-}
-// https://www.npmjs.com/package/react-hamburger-menu
-
-export default MenuMobile
